Rewrite site generator in Python
Hakyll has been a pain point for a while now: it's opinionated and
hard to change how it works; the templating system is terrible; and
the compile times are slow.  It's also over-complicated because I
don't need any of the caching or live serving stuff it does, I build
the site from scratch on every deploy.

I considered switching to a different site generator, but they will
all have those problems to some extent.  So instead just implement a
basic site generator myself.  This should be maximally flexible, and
the code isn't much longer than the hakyll-based solution.

One possible downside to this approach is that, by losing access to
pandoc's full API, I've had to resort to parsing the HTML to extract
the headings.  There may be other pain points discovered in the
future.  But that could be remedied by writing a small pandoc wrapper
in Haskell and building a container with that and the other website
dependencies installed.