Check publication date isn't after modification date
When the "published" field isn't set, a value is fetched automatically
using the git history.

When the "modified" field isn't set, the "date" field is used and, if
that's also missing, the "published" field is used.

This means that if something can have a published time which
is *before* the modified time, if the metadata refers to a time in the
past.  There are some memos which were imported from elsewhere, and so
the git creation date is after their actual creation date, so they
display as having a weirdly recent publication time.