Disable pandoc-filter-transform-code
I upgraded to pandoc-2.10, which panflute doesn't work with.  Through
silliness I garbage-collected the old docker image once I saw it was
building properly, I didn't actually try compiling the memo site...

Bringing back the old docker image is tricky because it references an
older version of texlive, which is 404ing, so given that the panflute
devs said (yesterday!) that the goal is to add support for pandoc-2.11
"within the next couple of days", I think it's best to just be
slightly broken for now and fix later.