Re-do CSS & layout
Largely similar in layout to what was before but, I hope, looks rather
nicer.  It's more mobile friendly, with breaks at 800px and 600px
width: shifting to single-column memos and a narrower sidebar on
listing pages at 800px, and single-column lists at 600px.

<table>s and <pre>s can overflow the content area, but they don't
stretch it.  Sidenotes are pretty bad, not sure what to do about those
really - they're readable, but it's not great.

The stylesheet is based somewhat on lookwhattheshoggothdraggedin.com.