Implement grid-based weeknote listing
Since these are the most-published type of memo, and they're strongly
linked to the calendar, I wanted a "calendar-style" view.  But a
calendar which has every day but Sunday blank is a lot of wasted
space, so instead I went for this grid-style approach which feels
similar in spirit to me, but has no gaps.