Render footnotes as footnotes, not sidenotes
Sidenotes don't work so well on a portrait monitor: either the notes
need horizontal scrolling, or the content column is too narrow.  So just
render them as footnotes, as they originally were.

This does mean you need to scroll down to see the notes, but since *I*
had to scroll anyway, and I am the primary audience of my memos, it's
not really a problem.

A bit of a hack is needed to make the references work with the `<base/>`
tag, so I should look into removing that some time.