Only include `h2`s in default toc
The memo template doesn't render a deeper toc, since it's unwieldy for
many memos.  I think what I really want is to be able to (optionally!)
override the toc in the frontmatter, support nesting in the template,
and so manually define a nested toc for those few memos where it's
useful.