Do not cache toc
Since metadata isn't included in the hash, changes to the
explicitly-given toc aren't reflected without manually deleting the
cache file.

Not an issue for memo.barrucadu.co.uk, since the pipeline regenerates
the site in full on every deploy, but it is a pain for local
development.

I considered caching a toc-from-the-html, but that didn't really make a
difference, so have just stuck with the simpler approach of caching only
the HTML.